% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{abort_on_error}
\alias{abort_on_error}
\title{Abort with a nice error message if there is an error in Rust.}
\usage{
abort_on_error(result)
}
\description{
Abort with a nice error message if there is an error in Rust.
}
\keyword{internal}
