% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_scenario.R
\name{add_households}
\alias{add_households}
\title{Add households to particular Census tracts in the region}
\usage{
add_households(scenario, tract, hhs)
}
\arguments{
\item{scenario}{An existing land-use scenario}

\item{tract}{a Census tract ID (or vector of tract IDs)-should be strings, not numbers}

\item{hhs}{Households to add to the tract}
}
\value{
a copy of the scenario with the requested households added

The hhs parameter should be a table/tibble/data.frame of households with the following columns:
\itemize{
\item income: Annual household income, dollars
\item hhsize: Household size
\item vehicles: Number of vehicles owned by the household
\item workers: Number of workers in the household
\item n: Number of households like this to add
}

If there are multiple tracts specified, the households will be evenly split among them
}
\description{
Add households to particular Census tracts in the region
}
