% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{all_or_nothing}
\alias{all_or_nothing}
\title{The all-or-nothing assignment step, where all vehicles traveling from A to B are assigned to the
same route. This is done repeatedly with new weights and averaged to distribute vehicles across
the network.}
\usage{
all_or_nothing(odflow_hourly, marginals, network, weights)
}
\description{
The all-or-nothing assignment step, where all vehicles traveling from A to B are assigned to the
same route. This is done repeatedly with new weights and averaged to distribute vehicles across
the network.
}
