% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{apply_direction_factors}
\alias{apply_direction_factors}
\title{This applies direction factors to an origin-destination matrix}
\usage{
apply_direction_factors(matrix, direction_factors)
}
\description{
This applies direction factors to an origin-destination matrix
}
