% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attraction.R
\name{balance_production_attraction}
\alias{balance_production_attraction}
\title{This adjusts the total attraction numbers to match total production. It returns a list
with members attractions containing the balanced attractions, production containing the
original productions, and factors containing the adjustment factors}
\usage{
balance_production_attraction(production, attraction)
}
\description{
This adjusts the total attraction numbers to match total production. It returns a list
with members attractions containing the balanced attractions, production containing the
original productions, and factors containing the adjustment factors
}
