% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf4sm-package.R
\docType{package}
\name{bf4sm-package}
\alias{bf4sm}
\alias{bf4sm-package}
\title{bf4sm: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
