% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{build_network}
\alias{build_network}
\title{Build the OSM network (using Julia)}
\usage{
build_network(osm, highway_types, installJulia = T)
}
\arguments{
\item{osm}{The OSM file to use}

\item{highway_types}{Character vector of highway=* tags to retain}

\item{installJulia}{Whether to install Julia if it is not found (default TRUE)}
}
\value{
A list with members $network (igraph network) and $network_geo (sf DataFrame with geographic outline of network)
}
\description{
Build the OSM network (using Julia)
}
