% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{calculate_direction_factors}
\alias{calculate_direction_factors}
\title{Up through the mode choice step, everything is done with productions at home and attractions at work.
For assignment, we need origins and destinations.
This calculates factors to figure out how many trips of each home-based type travel in each direction
during a particular time period}
\usage{
calculate_direction_factors(nhts)
}
\description{
Up through the mode choice step, everything is done with productions at home and attractions at work.
For assignment, we need origins and destinations.
This calculates factors to figure out how many trips of each home-based type travel in each direction
during a particular time period
}
