% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{calculate_occupancy_factors}
\alias{calculate_occupancy_factors}
\title{This calculates vehicle occupancy factors by trip type and time period}
\usage{
calculate_occupancy_factors(nhts)
}
\description{
This calculates vehicle occupancy factors by trip type and time period
}
\keyword{internal}
