% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{calibrate_trip_distance_betas}
\alias{calibrate_trip_distance_betas}
\title{Calibrate betas for all trip types
median_distances should be a list with element HBW, HBO, and NHB with median crow-flies
trip distances in each.}
\usage{
calibrate_trip_distance_betas(balanced, marginals, median_distances_m)
}
\description{
Calibrate betas for all trip types
median_distances should be a list with element HBW, HBO, and NHB with median crow-flies
trip distances in each.
}
