% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attraction.R
\name{estimate_attraction_functions}
\alias{estimate_attraction_functions}
\title{This estimates the distribution models on the Seattle data
It expects LODES data already aggregated to tracts}
\usage{
estimate_attraction_functions(trips, lodes)
}
\description{
This estimates the distribution models on the Seattle data
It expects LODES data already aggregated to tracts
}
