% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{estimate_median_crow_flies_distance}
\alias{estimate_median_crow_flies_distance}
\title{Estimates the median crow-flies distance from the NHTS in kilometers, by dividing the horse-flies distance
by 1.3. This is based on the paper https://arxiv.org/abs/2406.06490}
\usage{
estimate_median_crow_flies_distance(nhts)
}
\description{
Estimates the median crow-flies distance from the NHTS in kilometers, by dividing the horse-flies distance
by 1.3. This is based on the paper https://arxiv.org/abs/2406.06490
}
