% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modechoice.R
\name{estimate_mode_choice_model}
\alias{estimate_mode_choice_model}
\title{This estimates mode choice models based on the NHTS
This is much simplified from the traditional mode choice models used in demand modeling,
because we don't have attributes of the alternatives (or enough information to create them).
So we just use demographics and mode-invariant trip characteristics to estimate mode choice.}
\usage{
estimate_mode_choice_model(nhts)
}
\description{
This estimates mode choice models based on the NHTS
This is much simplified from the traditional mode choice models used in demand modeling,
because we don't have attributes of the alternatives (or enough information to create them).
So we just use demographics and mode-invariant trip characteristics to estimate mode choice.
}
