% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{estimate_production_functions}
\alias{estimate_production_functions}
\title{This estimates 20 trip production regressions - home based work (from home), home based work (to home),
home based other (from home), home based other (to home), and non-home-based, stratified by time of day
(AM peak, PM peak, midday, overnight).}
\usage{
estimate_production_functions(nhts)
}
\description{
Note that everything is ultimately joined to the households table; if you wish to filter households, you
can filter that table and not any of the others.
}
