% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{estimate_vmt}
\alias{estimate_vmt}
\title{Calculates VMT based on flows.}
\usage{
estimate_vmt(model, network, link_flows, period)
}
\arguments{
\item{model}{The estimated model object}

\item{network}{The network to use (should be the same one used in \code{\link[=network_assignment]{network_assignment()}}}

\item{link_flows}{Estimated link flows from the network assignment function}
}
\description{
Note that this is total VMT for the period the flows are for;
you would have to do this for each period to get total daily VMT.
}
