% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modechoice.R
\name{flow_by_mode}
\alias{flow_by_mode}
\title{Estimate numbers of trips by purpose and mode}
\usage{
flow_by_mode(odmat, marginals, mode_choice_models)
}
\description{
Estimate numbers of trips by purpose and mode
}
