% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{frank_wolfe}
\alias{frank_wolfe}
\title{This does the network assignment
The default relgap_tol is 1\% instead of the recommended 0.01\% to speed convergence in the
teaching environment.}
\usage{
frank_wolfe(
  odflow_hourly,
  marginals,
  network,
  maxiter = 100,
  relgap_tol = 0.01
)
}
\description{
This does the network assignment
The default relgap_tol is 1\% instead of the recommended 0.01\% to speed convergence in the
teaching environment.
}
