% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{get_aggregate_cost}
\alias{get_aggregate_cost}
\title{Get the aggregate cost to all travelers of the current assignment}
\usage{
get_aggregate_cost(network, flows)
}
\description{
Get the aggregate cost to all travelers of the current assignment
}
