% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attraction.R
\name{get_attraction_counts}
\alias{get_attraction_counts}
\title{This generates attraction estimates from LODES data
Marginals are only used to get the list of tracts, since
LODES does not contain all tracts}
\usage{
get_attraction_counts(marginals, attraction_functions)
}
\description{
This generates attraction estimates from LODES data
Marginals are only used to get the list of tracts, since
LODES does not contain all tracts
}
