% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_marginals.R
\name{get_base_marginals}
\alias{get_base_marginals}
\title{This gets baseline marginals for a region, from the 5-year ACS}
\usage{
get_base_marginals(state, county = NULL, year = NULL)
}
\description{
This gets baseline marginals for a region, from the 5-year ACS
}
