% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{get_distance_matrix}
\alias{get_distance_matrix}
\title{We use a singly-constrained (at origin) gravity model. For simplicity, we use straight-line
(great circle) distance between points.
This function estimates that distance matrix.}
\usage{
get_distance_matrix(marginals)
}
\description{
We use a singly-constrained (at origin) gravity model. For simplicity, we use straight-line
(great circle) distance between points.
This function estimates that distance matrix.
}
