% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{get_flows}
\alias{get_flows}
\title{Perform the singly-constrained gravity model to get zone-to-zone flows based on balanced
trip tables and estimated betas from above.}
\usage{
get_flows(balanced, marginals, betas)
}
\description{
Perform the singly-constrained gravity model to get zone-to-zone flows based on balanced
trip tables and estimated betas from above.
}
