% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{get_hh_counts}
\alias{get_hh_counts}
\title{This generates the number of households in each category in a geographic area
given marginal distributions.}
\usage{
get_hh_counts(marginals, seed)
}
\description{
The marginal distributions we expect are
\itemize{
\item Household size (hhsize), topcoded at HHSIZE_TOPCODE
\item Vehicles available (vehicles), topcoded at VEHICLES_TOPCODE
\item Number of workers (workers), topcoded at WORKERS_TOPCODE
\item Income category (income), see load_nhts.R
}
}
\details{
These should be formatted as a table with columns geoid, marginal, value, count,
where geoid is a tract ID, marginal is the name of one of the marginals (above),
value is the value for that marginal (e.g. vehicles = 2 for two-vehicle households),
and count is the expected count in that category
}
