% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{get_hh_counts_for_tract}
\alias{get_hh_counts_for_tract}
\title{This generates the number of households in each category for a specific geographic area (e.g. tract).}
\usage{
get_hh_counts_for_tract(grp, id, seed)
}
\description{
This generates the number of households in each category for a specific geographic area (e.g. tract).
}
