% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{get_lane_capacity}
\alias{get_lane_capacity}
\title{Gets lane capacity based on OSM highway type
Capacities come from Mannering and Washburn, Principles of Highway Engineering and Traffic Analysis, 7th ed.}
\usage{
get_lane_capacity(highway_types)
}
\description{
Gets lane capacity based on OSM highway type
Capacities come from Mannering and Washburn, Principles of Highway Engineering and Traffic Analysis, 7th ed.
}
\keyword{internal}
