% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{get_nhts_trip_counts}
\alias{get_nhts_trip_counts}
\title{Get the trip counts in each category and time period for each household}
\usage{
get_nhts_trip_counts(trips)
}
\description{
Get the trip counts in each category and time period for each household
}
