% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{get_production_counts}
\alias{get_production_counts}
\title{This generates trip counts by tract and time period, based on marginals and
trip regressions. It expects marginals in the format described in the docs for
get_hh_counts, and areas with a geoid and area_sqmi column with the area of each
geographic unit}
\usage{
get_production_counts(marginals, generation_functions, seed)
}
\description{
This generates trip counts by tract and time period, based on marginals and
trip regressions. It expects marginals in the format described in the docs for
get_hh_counts, and areas with a geoid and area_sqmi column with the area of each
geographic unit
}
