% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_periods.R
\name{get_time_period}
\alias{get_time_period}
\title{Get the time period from an NHTS time.
Currently these are hardcoded as AM Peak 6-10 AM, Midday 10-4, PM Peak 4-7, Overnight 7PM - 6AM.}
\usage{
get_time_period(time)
}
\description{
Get the time period from an NHTS time.
Currently these are hardcoded as AM Peak 6-10 AM, Midday 10-4, PM Peak 4-7, Overnight 7PM - 6AM.
}
