% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{get_trip_type}
\alias{get_trip_type}
\title{This returns the trip type for each trip in an NHTS trip table. It will classify trips as
HBW - Home based work
HBO - Home based other
NHB - Non-home-based}
\usage{
get_trip_type(whyfrom, whyto)
}
\description{
This returns the trip type for each trip in an NHTS trip table. It will classify trips as
HBW - Home based work
HBO - Home based other
NHB - Non-home-based
}
