% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{ipf}
\alias{ipf}
\title{We implement IPF on our own in base R and Rust.}
\usage{
ipf(seed, marginals)
}
\arguments{
\item{seed}{should be a data frame with a row for each household, and a column for each marginal
indicating what value that household takes on (e.g. age, race, etc columns). It should also have a weight column indicating
the seed weight for that household (or type of household).}

\item{marginals}{should be a data frame with a row for each marginal/value combination, and
a column \code{marginal} indicating which marginal this is (should match column names in \code{seed}),
a column \code{value} indicating what value of the marginal this is (should match column values in \code{seed}),
and a column \code{count} indicating the target households in this marginal/value combination.}
}
\description{
We implement IPF on our own in base R and Rust.
}
\keyword{internal}
