% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_marginals.R
\name{load_landuse_scenario}
\alias{load_landuse_scenario}
\title{Load a land use scenario in Excel format}
\usage{
load_landuse_scenario(filename)
}
\description{
Load a land use scenario in Excel format
}
