% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{load_model_v0}
\alias{load_model_v0}
\title{Load a model in model format 0 (used with pre-2026 releases)}
\usage{
load_model_v0(filename)
}
\arguments{
\item{filename}{Filename or URL for model}
}
\description{
Load a model in model format 0 (used with pre-2026 releases)
}
