% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nhts.R
\name{load_nhts}
\alias{load_nhts}
\title{Load 2017 NHTS data, handling types appropriately}
\usage{
load_nhts(base_path)
}
\arguments{
\item{base_path}{path to folder containing 2017 NHTS CSV files}
}
\description{
Load 2017 NHTS data, handling types appropriately
}
