% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_congestion}
\alias{map_congestion}
\title{Produce a congestion map}
\usage{
map_congestion(model, network, flows)
}
\arguments{
\item{model}{the model in use}

\item{network}{the network scenario in use (e.g. model$networks$baseline)}

\item{flows}{output of \code{\link[=network_assignment]{network_assignment()}}}
}
\description{
Produce a congestion map
}
