% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_trip_distribution}
\alias{map_trip_distribution}
\title{Map trip distribution from a single trip.}
\usage{
map_trip_distribution(model, flows, timeperiod, triptype, origin_tract)
}
\arguments{
\item{model}{the model to use}

\item{timeperiod}{Time period to map (AM Peak, Midday, PM Peak, Overnight)}

\item{triptype}{Trip type to map (HBO, HBW, NHB)}

\item{origin_tract}{Tract to map productions of}

\item{trip_count}{\code{\link[=trip_generation]{trip_generation()}} results}
}
\description{
Produces a map showing where trips produced in \code{origin_tract} are attracted.
}
