% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_trip_generation}
\alias{map_trip_generation}
\title{Map trip generation.}
\usage{
map_trip_generation(model, trip_counts, end, timeperiod, triptype)
}
\arguments{
\item{model}{the model to use}

\item{end}{"productions" or "attractions"}

\item{timeperiod}{Time period to map (AM Peak, Midday, PM Peak, Overnight)}

\item{triptype}{Trip type to map (HBO, HBW, NHB)}

\item{trip_count}{\code{\link[=trip_generation]{trip_generation()}} results}
}
\description{
Produces a map showing where trips are produced or attracted.
}
