% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{mode_choice}
\alias{mode_choice}
\title{This runs the mode choice step of the model, and returns flows differentiated
by mode for each trip type and time of day.}
\usage{
mode_choice(model, marginals, flows)
}
\description{
This runs the mode choice step of the model, and returns flows differentiated
by mode for each trip type and time of day.
}
