% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFirstFourStepModel-package.R
\docType{package}
\name{MyFirstFourStepModel-package}
\alias{MyFirstFourStepModel}
\alias{MyFirstFourStepModel-package}
\title{MyFirstFourStepModel: Highly Simplified Four-Step Travel Demand Models for Teaching}
\description{
This package provides a highly simplified four-step model. Running the model requires only R, and is designed to be computationally efficient and run even on consumer-grade equipment.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://projects.indicatrix.org/MyFirstFourStepModel/}
}

}
\author{
\strong{Maintainer}: Matt Bhagat-Conway \email{mwbc@unc.edu} (\href{https://orcid.org/0000-0002-1210-2982}{ORCID})

}
\keyword{internal}
