% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{network_assignment}
\alias{network_assignment}
\title{This runs the network assignment step of the model.}
\usage{
network_assignment(model, scenario, network, mode_flows, period)
}
\arguments{
\item{model}{The estimated model}

\item{scenario}{The scenario to use}

\item{mode_flows}{Flows by mode and time of day, output of mode_choice function}

\item{period}{Time period to assign, can be "AM Peak", "Midday", "PM Peak", or "Overnight".}
}
\description{
Network assignment is based on a Frank-Wolfe static traffic assignment algorithm.
This returns a list of link-level flows. It applies peaking factors and average occupancy
before doing assignment to convert period person-trips to hourly vehicle-trips.
}
