% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{save_model}
\alias{save_model}
\title{Save a model}
\usage{
save_model(model, filename)
}
\arguments{
\item{model}{model to save}

\item{filename}{file to save to (canonical extension .mf4sm)}
}
\description{
Save a model
}
