% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_simple.R
\name{write_mnl}
\alias{write_mnl}
\title{Write just enough of a multinomial logit model that we can deserialize and apply it}
\usage{
write_mnl(model, archive, name)
}
\description{
Write just enough of a multinomial logit model that we can deserialize and apply it
}
