% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{estimate}
\alias{estimate}
\title{Estimate a four step model for later use, based on 2017 NHTS data and PSRC household
survey data (for distribution functions).}
\usage{
estimate(
  nhts,
  osm,
  state,
  county,
  year,
  highway_types = c("motorway", "motorway_link", "trunk", "trunk_link", "primary",
    "primary_link"),
  installJulia = T
)
}
\arguments{
\item{nhts}{Path to 2017 NHTS CSV files}

\item{osm}{Path to OSM .pbf file}

\item{state}{State to estimate model for}

\item{county}{County (or vector of counties) to estimate model for}

\item{year}{Year of ACS and LODES data to use (if you get 404 errors, you are probably trying to use a year that LODES is not available for)}

\item{highway_types}{OSM highway= tags to include in network, default "motorway", "motorway_link", "trunk", "trunk_link", "primary", "primary_link"}

\item{installJulia}{Install Julia for network building if it is not found (default true)}
}
\description{
Load the NHTS data with load_nhts(), and if desired filter the households table to just the households
you want to use in estimation.
}
