% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{estimate}
\alias{estimate}
\title{Estimate a four step model for later use}
\usage{
estimate(
  nhts,
  seed_matrix,
  psrc,
  psrc_lodes,
  base_marginals,
  network,
  network_geo
)
}
\description{
Estimate a four step model for later use
}
