% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{estimate_vmt}
\alias{estimate_vmt}
\title{This function calculates VMT based on flows. Note that this is total VMT for the period the flows are for;
you would have to do this for each period to get total daily VMT.}
\usage{
estimate_vmt(model, marginals, link_flows, period)
}
\description{
This function calculates VMT based on flows. Note that this is total VMT for the period the flows are for;
you would have to do this for each period to get total daily VMT.
}
