% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modechoice.R
\name{get_mode_shares}
\alias{get_mode_shares}
\title{This function uses the output of \code{\link[=flow_by_mode]{flow_by_mode()}} to calculate mode shares, returning a data frame with
rows for car, bike, walk, and transit.}
\usage{
get_mode_shares(flows_by_mode)
}
\description{
This function uses the output of \code{\link[=flow_by_mode]{flow_by_mode()}} to calculate mode shares, returning a data frame with
rows for car, bike, walk, and transit.
}
