% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{network_to_gis}
\alias{network_to_gis}
\title{Convenience function to export a network scenario to a GIS file}
\usage{
network_to_gis(network, file)
}
\description{
Convenience function to export a network scenario to a GIS file
}
