% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{trip_distribution}
\alias{trip_distribution}
\title{This runs the trip distribution step of the model}
\usage{
trip_distribution(model, marginals, balanced)
}
\description{
Trip distribution is based on the relative locations of the tracts (in \code{marginals}),
the betas estimated during the estimation phase of the model, and the total trip productions
and attractions.
}
