% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{trip_generation}
\alias{trip_generation}
\title{This runs the trip generation step of the model.}
\usage{
trip_generation(model, marginals)
}
\description{
Trip production is based on the marginal distributions for vehicles, workers, household size, and
income. Trip attraction is based on job counts. Attractions are balanced to match productions before
being returned.
}
