% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_simple.R
\name{write_lm}
\alias{write_lm}
\title{Write the minimal information to be able to reconstruct enough of an
lm to be able to do prediction.}
\usage{
write_lm(model, archive, name)
}
\description{
Write the minimal information to be able to reconstruct enough of an
lm to be able to do prediction.
}
